backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
openai==1.99.9
faiss-cpu==1.12.0
numpy==2.3.2
pdfplumber==0.11.7
tiktoken==0.9.0
pydantic==2.11.7
python-multipart==0.0.20
